requirements.txt
tensorflow>=2.0
keras>=2.0
pillow
matplotlib
numpy